added notification form, fixed xml, clean of DI, added serilog, added segmented sieve prime generator for large primes

+++ PrimeNumbers.BLL/Services/Interfaces/IPrimesGenerator.cs
@@ -7,6 +7,6 @@ namespace PrimeNumbers.BLL.Services.Interfaces
     {
         bool TurnedOn { get; set; }
 
-        Task<(IList<int>, int, int, int)> GenerateUsingSieveOfAtkin(int limit, long elpsd, int x = 1, int y = 1, int lastPrime = 1);
+        Task<(IList<long>, int, int, int)> GenerateUsingSieveOfAtkin(int limit, long elpsd, int x = 1, int y = 1, int lastPrime = 1);
     }
 }
