+++ PrimeNumbers.DTO/PrimeGenerationState.cs
@@ -6,7 +6,7 @@ namespace PrimeNumbers.DTO
 {
     public class PrimeGenerationState
     {
-        public int Limit { get; set; }
+        public long Limit { get; set; }
         public int X { get; set; }
         public int Y { get; set; }
     }
